+++ Assets/Scripts/BuildMode/RoadCreateState.cs
@@ -16,7 +16,7 @@ public class RoadCreateState : IBuildingState
     float length;
     private List<Vector3> posList;
     private Vector3 displayPosition; 
-    int rotation;
+    float rotation;
 
     public RoadCreateState(Vector3 gridPosition,
                            int iD,
@@ -65,7 +65,7 @@ public class RoadCreateState : IBuildingState
         }
     }
 
-    public void OnModify(Vector3 gridPosition, int rotation = 0)
+    public void OnModify(Vector3 gridPosition, float rotation = 0)
     {
         if (previewSystem.expand == true)
         {
@@ -153,7 +153,7 @@ public class RoadCreateState : IBuildingState
         return true;
     }
 
-    public void UpdateState(Vector3 gridPosition, int rotation = 0)
+    public void UpdateState(Vector3 gridPosition, float rotation = 0)
     {
         bool placementValidity = CheckPlacementValidity(gridPosition, selectedObjectIndex);
