+++ Assets/Scripts/BuildMode/WallCreateState.cs
@@ -46,7 +46,7 @@ public class WallCreateState : IBuildingState
     }
 
 
-    public void OnModify(Vector3 gridPosition, int rotation = 0)
+    public void OnModify(Vector3 gridPosition, float rotation = 0)
     {
         grid = placementSystem.GetCurrentGrid();
         wallMapping = placementSystem.GetCurrentWalls();
@@ -139,7 +139,7 @@ public class WallCreateState : IBuildingState
     }
 
 
-    public void UpdateState(Vector3 gridPosition, int rotation = 0)
+    public void UpdateState(Vector3 gridPosition, float rotation = 0)
     {
         bool placementValidity = CheckPlacementValidity(gridPosition);
